+++ dmdownloader/downloader/gui.py
@@ -3,6 +3,7 @@ import tkinter.messagebox as msg
 import tkinter as tk
 import dmdownloader.downloader.asyntk as asyntk
 import dmdownloader.downloader.sitelib.service as service
+import dmdownloader.downloader.components as cmp
 
 FONT_NAME = "微软雅黑"
 
@@ -45,8 +46,7 @@ class DownloaderApp(tk.Tk):
         self.frames["setting_frame"] = InitFrame(master=container, controller=self)
         self.frames["waiting_frame"] = WaitingFrame(master=container, controller=self)
         self.frames["main_frame"] = VetcScrollFrame(lambda master: MainFrame(master=master, favorites=favorites, controller=self), 800, height, master=container)
-        self.frames["setting_frame"].cookie_eny.insert(0, self.app_config["cookie"])
-        self.frames["setting_frame"].ua_eny.insert(0, self.app_config["user_agent"])
+        self.frames["setting_frame"].load_config()
 
         self.frames["main_frame"].grid(row=0, column=0, sticky="nsew")
         self.frames["setting_frame"].grid(row=0, column=0, sticky="nsew")
@@ -56,7 +56,7 @@ class DownloaderApp(tk.Tk):
         
         self.asyn_event = asyntk.AsyncEvent(self)
 
-        self.show_frame("setting_frame")
+        self.show_frame("main_frame")
 
     def show_frame(self, page_name):
         """ 切换页面 """
@@ -89,34 +89,66 @@ class InitFrame(ttk.Frame):
         super().__init__(master, relief="sunken")
         self.master = master
         self.controller = controller
+        self.global_config = controller.app_config
 
-        self.get_input_frame().pack(ipady=20)
-        self.get_btn_frame().pack(ipady=20)
+        ttk.Label(self, text="设置", style="title.TLabel").pack(ipady=20)
+        self.get_input_frame().pack()
+        self.get_btn_frame().pack()
     
     def get_input_frame(self):
         fra = ttk.Frame(self)
         
-        ua = tk.StringVar()
-        cookie = tk.StringVar()
-        self.ua_eny = ttk.Entry(fra, width=75, font=(FONT_NAME, 10), textvariable=ua)
-        self.cookie_eny = ttk.Entry(fra, width=75, font=(FONT_NAME, 10), textvariable=cookie)
+        self.ua_var = tk.StringVar()
+        self.cookie_var = tk.StringVar()
+        self.top_filter_var = tk.BooleanVar()
+        self.bottom_filter_var = tk.BooleanVar()
+        self.zhconv_var = tk.BooleanVar()
+        self.raw_var = tk.BooleanVar()
 
-        ttk.Label(master=fra, text="\n\n\nUserAgent:", width=50, style="ep_title.TLabel").pack()
-        self.ua_eny.pack()
-        ttk.Label(master=fra, text="Cookie:", width=50, style="ep_title.TLabel").pack()
-        self.cookie_eny.pack()
-        
+        cmp.inputBox(fra, "UserAgent:", self.ua_var, (FONT_NAME, 10), "ep_title.TLabel").pack()
+        cmp.inputBox(fra, "Cookie:", self.cookie_var, (FONT_NAME, 10), "ep_title.TLabel").pack()
+
+        opt_fra = ttk.Frame(fra)
+        fra_top, self.opt_top = cmp.optBox(opt_fra, "顶部弹幕过滤", self.top_filter_var, "ep_title.TLabel")
+        fra_top.pack(side="left", padx=30)
+        fra_bot, self.opt_bot = cmp.optBox(opt_fra, "底部弹幕过滤", self.bottom_filter_var, "ep_title.TLabel")
+        fra_bot.pack(side="left", padx=30)
+        fra_zh, self.opt_zh = cmp.optBox(opt_fra, "繁转换", self.zhconv_var, "ep_title.TLabel")
+        fra_zh.pack(side="left", padx=30)
+        fra_raw, self.opt_raw = cmp.optBox(opt_fra, "下载源文件", self.raw_var, "ep_title.TLabel")
+        fra_raw.pack(side="left", padx=30)
+        opt_fra.pack()
+
         return fra
+    
+    def load_config(self):
+        config = self.global_config
+        self.ua_var.set(config["user_agent"])
+        self.cookie_var.set(config["cookie"])
+
+        self.opt_top.set(int(config["top_filter"]))
+        self.opt_bot.set(int(config["bottom_filter"]))
+        self.opt_zh.set(int(config["open_zhconv"]))
+        self.opt_raw.set(int(config["download_raw"]))
+        self.top_filter_var.set(config["top_filter"])
+        self.bottom_filter_var.set(config["bottom_filter"])
+        self.zhconv_var.set(config["open_zhconv"])
+        self.raw_var.set(config["download_raw"])
+    
+    def save_config(self):
+            config = self.global_config
+            self.global_config["user_agent"] = self.ua_var.get()
+            self.global_config["cookie"] = self.cookie_var.get()
+            config["top_filter"] = self.top_filter_var.get()
+            config["bottom_filter"] = self.bottom_filter_var.get()
+            config["open_zhconv"] = self.zhconv_var.get()
     
     def get_btn_frame(self):
         fra = ttk.Frame(self)
-
         def handler():
-            self.controller.app_config["user_agent"] = self.ua_eny.get()
-            self.controller.app_config["cookie"] = self.cookie_eny.get()
+            self.save_config()
             self.controller.show_frame("main_frame")
         ttk.Button(fra, text="确认", style="favortes.TButton", command=handler).pack()
-
         return fra
 
 class MainFrame(ttk.Frame):